+++ __tests__/inpn-proxy.test.js
@@ -1,25 +1,4 @@
-const fs = require('fs');
-const vm = require('vm');
-
-function loadHandler(mockFetch) {
-  const code = fs.readFileSync('netlify/functions/inpn-proxy.js', 'utf-8');
-  const patched = code.replace(
-    /const fetch = \(\.\.\.args\) => import\("node-fetch"\)\.then\(\(\{default: f\}\) => f\(\.\.\.args\)\);/,
-    'const fetch = (...args) => global.__fetch(...args);'
-  );
-  const context = { require, console, exports: {}, __fetch: mockFetch };
-  context.global = context;
-  vm.createContext(context);
-  vm.runInContext(patched, context);
-  return context.exports.handler;
-}
-
-function mockFetch(html) {
-  return jest.fn().mockResolvedValue({
-    ok: true,
-    text: () => Promise.resolve(html)
-  });
-}
+const { loadHandler, mockFetch } = require('../test-utils');
 
 describe('inpn-proxy handler', () => {
   test('extracts canvas for carte type', async () => {
@@ -37,5 +16,26 @@ describe('inpn-proxy handler', () => {
     const handler = loadHandler(fetchMock);
     const res = await handler({ queryStringParameters: { cd: '1', type: 'statut' } });
     expect(res.statusCode).toBe(404);
+  });
+
+  test('returns 400 for unsupported type', async () => {
+    const fetchMock = mockFetch('');
+    const handler = loadHandler(fetchMock);
+    const res = await handler({ queryStringParameters: { cd: '1', type: 'foo' } });
+    expect(res.statusCode).toBe(400);
+  });
+
+  test('returns 400 when cd is missing', async () => {
+    const fetchMock = mockFetch('');
+    const handler = loadHandler(fetchMock);
+    const res = await handler({ queryStringParameters: { type: 'carte' } });
+    expect(res.statusCode).toBe(400);
+  });
+
+  test('returns 400 when type parameter missing', async () => {
+    const fetchMock = mockFetch('');
+    const handler = loadHandler(fetchMock);
+    const res = await handler({ queryStringParameters: {} });
+    expect(res.statusCode).toBe(400);
   });
 });
